released sr2019

+++ scripts/make-src-release.sh
@@ -1,24 +1,34 @@
 #!/bin/sh
+die () {
+  echo "make-src-release.sh: error: $*" 1>&2
+  exit 1
+}
 cd `dirname $0`/..
-root=`pwd`
-version="`cat VERSION`"
-gitid="`./scripts/get-git-id.sh|sed -e 's,^\(.......\).*,\1,'`"
+version="`awk '/define VERSION/{print $4}' src/version.cpp|sed -e 's,",,g'`"
+VERSION="`cat VERSION`"
+[ "$version" = "$VERSION" ] || \
+die "versions '$version' in 'src/version.cpp' and " \
+    "'$VERSION' in 'VERSION' do not match"
+fullgitid="`./scripts/get-git-id.sh`"
+gitid="`echo $fullgitid|sed -e 's,^\(.......\).*,\1,'`"
+branch=`git branch|grep '^\*'|head -1|awk '{print $2}'`
+[ "$branch" = "" ] && die "could not get branch"
 name=cadical-${version}-${gitid}
+[ "$branch" = "master" ] || name="$name-$branch"
 dir=/tmp/$name
 tar=/tmp/$name.tar.xz
 rm -rf $dir
 mkdir $dir || exit 1
-cd $dir || exit 1
-git clone $root $dir || exit 1
-rm -rf TODO.md test $dir/scripts
-sed -i -e '/rm -f test/d' $dir/makefile.in
-sed -i -e '/optionally test it/d' $dir/configure
-mkdir $dir/scripts
-sed \
-  -e 's,`../scripts/get-git-id.sh`,'"`$root/scripts/get-git-id.sh`", \
-  $root/scripts/make-config-header.sh > $dir/scripts/make-config-header.sh
+git archive $branch | tar -x -C $dir
+cat >$dir/scripts/get-git-id.sh <<EOF
+#!/bin/sh
+echo "$fullgitid"
+EOF
+sed -i -e "s,IDENTIFIER 0$,IDENTIFIER \"$fullgitid\"," $dir/src/version.cpp
 chmod 755 $dir/scripts/*.sh
 cd /tmp
 rm -rf /tmp/$name/.git
 tar cJf $tar $name
-ls -l $tar
+bytes="`ls --block-size=1 -s $tar 2>/dev/null |awk '{print $1}'`"
+echo "generated '$tar' of $bytes bytes"
+rm -rf $dir
